ref(onboarding): Split powershell onboarding doc into multiple files
