+++ static/app/gettingStartedDocs/powershell/powershell/index.tsx
@@ -1,0 +1,11 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+
+import {crashReport} from './crashReport';
+import {onboarding} from './onboarding';
+
+const docs: Docs = {
+  onboarding,
+  feedbackOnboardingCrashApi: crashReport,
+};
+
+export default docs;
