+++ static/app/gettingStartedDocs/powershell/powershell/onboarding.spec.tsx
@@ -2,7 +2,7 @@ import {renderWithOnboardingLayout} from 'sentry-test/onboarding/renderWithOnboa
 import {screen} from 'sentry-test/reactTestingLibrary';
 import {textWithMarkupMatcher} from 'sentry-test/utils';
 
-import docs from './powershell';
+import docs from '.';
 
 describe('powershell onboarding docs', () => {
   it('renders docs correctly', async () => {
